chore: update sentiment model to export single output and remove deprecated files
