--- onnx-tracer/models/sentiment_simple/vocab.json
@@ -1,15 +1,0 @@
-{
-  "i": 1,
-  "love": 2,
-  "this": 3,
-  "is": 4,
-  "great": 5,
-  "happy": 6,
-  "with": 7,
-  "the": 8,
-  "result": 9,
-  "hate": 10,
-  "bad": 11,
-  "not": 12,
-  "satisfied": 13
-}
